refactor: move util to Store as staticmethod
